+++ textblob_de/sentiments.py
@@ -37,6 +37,7 @@ Main resource for ``de-sentiment.xml``:
 from __future__ import absolute_import
 from textblob.base import BaseSentimentAnalyzer, CONTINUOUS
 from textblob_de.tokenizers import PatternTokenizer
+from textblob_de.lemmatizers import PatternParserLemmatizer
 from textblob_de.de import sentiment as pattern_sentiment
 
 
@@ -47,13 +48,38 @@ class PatternAnalyzer(BaseSentimentAnalyzer):
 
     ``(polarity, subjectivity)``
     '''
-    def __init__(self, tokenizer=None):
+    def __init__(self, tokenizer=None, lemmatizer=None, lemmatize=True):
         self.tokenizer = tokenizer if tokenizer else PatternTokenizer()
+        self.lemmatize = lemmatize if lemmatize else True
+        if self.lemmatize:
+            self.lemmatizer = lemmatizer if lemmatizer \
+                else PatternParserLemmatizer(tokenizer=self.tokenizer)
+        
     
     def analyze(self, text):
         """Return the sentiment as a tuple of the form:
         ``(polarity, subjectivity)``
         
         :param str text: A string.
+        
+        .. todo::
+        
+            Figure out best format to be passed to the analyzer.
+            There might be a better format than a string of space separated
+            lemmas (e.g. with pos tags) but the parsing/tagging
+            results look rather inaccurate and a wrong pos
+            might prevent the lexicon lookup of an otherwise correctly 
+            lemmatized word form (or would it not?) - further checks needed.
+            
         """
-        return pattern_sentiment(text, self.tokenizer)
+        if self.lemmatize:
+            text = self._lemmatize(text)
+        _sentiment = pattern_sentiment(text, self.tokenizer)
+        return _sentiment
+    
+    def _lemmatize(self, raw):
+        # returns a list of [(lemma1, tag1), (lemma2, tag2), ...]
+        _lemmas = self.lemmatizer.lemmatize(raw)
+        # pass to analyzer as a string
+        _lemmas = " ".join([l for l, t  in _lemmas])
+        return _lemmas